useful_time/requirements.txt
asgiref==3.5.1
Django==3.2
django-colorfield==0.6.3
django-currentuser==0.5.3
django-debug-toolbar==3.4.0
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
